requirements.txt
pydantic>=2.4.2
langchain-openai>=0.0.1
langchain-core>=0.1.0
youtube-transcript-api>=0.6.1
langgraph>=0.0.15
python-dotenv>=1.0.0
pytube>=15.0.0
requests>=2.31.0
typing-extensions>=4.5.0
beautifulsoup4>=4.12.0
